refactor(hv-dataqc): rename to valid Python package, use -m invocation
Rename hv-dataqc/ → hv_dataqc/ and subdirs (cache-fetcher, extract-source,
extract-harmonized) → underscore versions so they can be importable Python
packages. Add __init__.py files, drop the sys.path hack, and switch shell
script invocations to `python -m hv_dataqc.<subpkg>.<module>` from the repo
root. Also un-ignore the cache_fetcher/ source dir (only the fetched
output under local_output/ should be ignored).
