+++ .gitignore
@@ -8,5 +8,5 @@ pat.txt
 **/Output/
 
 # HV-DataQC working output (downloaded JSONs, dbgap-cache, comparison reports)
-hv-dataqc/sb_output/
-hv-dataqc/local_output/
+hv_dataqc/sb_output/
+hv_dataqc/local_output/
