--- hv-dataqc/.gitignore
@@ -1,2 +1,0 @@
-local_output
-cache-fetcher
