+++ hv_dataqc/compare/config/clinical_ranges.yaml
@@ -1,4 +1,4 @@
-# hv-dataqc clinical ranges configuration
+# hv_dataqc clinical ranges configuration
 # =========================================
 # Per-variable plausible ranges for flagging impossible/suspicious values.
 # Used by check_clinical_ranges.py and compare_source_harmonized.py.
