+++ hv_dataqc/compare/config/thresholds.yaml
@@ -1,4 +1,4 @@
-# hv-dataqc statistical comparison thresholds
+# hv_dataqc statistical comparison thresholds
 # =============================================
 # Controls PASS / WARN / FAIL boundaries for all statistical checks (C1-C8).
 # Clinical plausible-range bounds (C9) and cross-variable rules (C10) remain in
