+++ hv_dataqc/sb_scripts/setup.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 # Run once per SB session to install uv and project dependencies.
 pip install uv
-# Derive repo root from this script's location (sb_scripts/ -> hv-dataqc/ -> repo root)
+# Derive repo root from this script's location (sb_scripts/ -> hv_dataqc/ -> repo root)
 HV="$(cd "$(dirname "${BASH_SOURCE[0]}")/../.." && pwd)"
 cd "$HV"
 export UV_LINK_MODE=copy
